Modification necessary to specify drifters at IC
I added three functions as methods to the WaveVortexModel class. Also I added some properties (X, Y, Z) to the object created by the WaveVortexModel class.
I think these modifications are necessary for Specify floats/drifters and place drifters on an isopycnal under initial conditions. But I'm not sure about that or if it can be done any better.